docs/locust/Collection/locust_user/requirements.txt
locust
pyjwt